make the state transition base as an example implementaion
